feat: add confirmation deleteCheck

+++ app/components/CheckList.js
@@ -1,9 +1,9 @@
 'use client'
 
-import Link from 'next/link'
-
 import { useState, useEffect } from 'react'
+import Link from 'next/link'
 import useSWR, { mutate } from 'swr'
+import toast from 'react-hot-toast'
 
 import { fetcher } from '@/helpers/fetcher'
 import { formatDate } from '@/helpers/formatDate'
@@ -11,13 +11,16 @@ import { formatDate } from '@/helpers/formatDate'
 import downloadNotes from '@/helpers/downloadNotes'
 import Download from '@/public/download.svg'
 
-import toast from 'react-hot-toast'
 import Loader from '@/app/components/Loader'
-import { useTranslation } from '../i18n/client'
+import { useTranslation } from '@/app/i18n/client'
+import DeleteModal from '@/app/components/DeleteModal'
 
 const CheckList = ({ projectId, bookId, lng }) => {
   const { t } = useTranslation(lng, 'common')
+
   const [notesCounts, setNotesCounts] = useState({})
+  const [showDeleteModal, setShowDeleteModal] = useState(false)
+  const [checkToDelete, setCheckToDelete] = useState(null)
 
   const { data: checks, error } = useSWR(
     projectId && bookId && `/api/projects/${projectId}/books/${bookId}/checks`,
@@ -49,29 +52,41 @@ const CheckList = ({ projectId, bookId, lng }) => {
         document.body.removeChild(link)
       })
       .catch((error) => {
-        const message = t('errorDownloadNotes') + ' ' + error
+        const message = `${t('errorDownloadNotes')} ${error}`
         toast.error(message)
       })
   }
 
-  const handleDeleteCheck = async (checkId) => {
-    try {
-      const response = await fetch(
-        `/api/projects/${projectId}/books/${bookId}/checks/${checkId}`,
-        {
-          method: 'DELETE',
+  const openDeleteModal = (check) => {
+    setCheckToDelete(check)
+    setShowDeleteModal(true)
+  }
+
+  const confirmDeleteCheck = async () => {
+    if (checkToDelete) {
+      try {
+        const response = await fetch(
+          `/api/projects/${projectId}/books/${bookId}/checks/${checkToDelete.check_id}`,
+          {
+            method: 'DELETE',
+          }
+        )
+
+        if (!response.ok) {
+          throw new Error(t('errorDeletingCheck'))
         }
-      )
 
-      if (!response.ok) {
-        throw new Error(t('errorDeletingCheck'))
+        toast.success(t('checkDeleted'))
+        mutate(`/api/projects/${projectId}/books/${bookId}/checks`) // Обновляем данные после удаления
+      } catch (error) {
+        toast.error(t('errorOccurred'))
       }
-
-      toast.success(t('checkDeleted'))
-      mutate(`/api/projects/${projectId}/books/${bookId}/checks`)
-    } catch (error) {
-      toast.error(t('errorOccurred'))
     }
+    setShowDeleteModal(false)
+  }
+
+  const cancelDeleteCheck = () => {
+    setShowDeleteModal(false)
   }
 
   return (
@@ -126,8 +141,8 @@ const CheckList = ({ projectId, bookId, lng }) => {
                   </td>
                   <td className="border p-2 text-center">
                     <button
-                      onClick={() => handleDeleteCheck(check.check_id)}
                       className="text-red-600 hover:text-red-800"
+                      onClick={() => openDeleteModal(check)}
                     >
                       {t('delete')}
                     </button>
@@ -139,6 +154,15 @@ const CheckList = ({ projectId, bookId, lng }) => {
         </div>
       ) : (
         <Loader />
+      )}
+
+      {showDeleteModal && (
+        <DeleteModal
+          isVisible={showDeleteModal}
+          message={t('confirmDeleteCheck')}
+          onConfirm={confirmDeleteCheck}
+          onCancel={cancelDeleteCheck}
+        />
       )}
     </>
   )
